[iOS] Refactor: 1. UseCase객체를 만들어서 CategoryViewModel이라는 	유의미한 모델 객체를 핸들러에 던지도록 기능 수정함 	2. 함수가 하나의 책임만을 하도록 함수를 분리함 	   => url배열을 일단 request 해서 url 배열을 받으면 	      url배열을 가지고 usecase를 설정하는 메소드를 호출하도록 함
